+++ packages/react-native-core/src/internal/Callback.ts
@@ -1,9 +1,22 @@
-import { NativeError } from './NativeError'
+import { NativeError } from './NativeError';
 
 export type Callback<T> = (
   error: NativeError | undefined,
   result: T | undefined,
 ) => void;
+
+export function mergeVoid<Error>(
+  mapError: (error: NativeError) => Error,
+  errorCallback: (error: Error) => void,
+  successCallback: () => void,
+): Callback<void> {
+  return (error, result) => {
+    if (error) {
+      errorCallback(mapError(error));
+    }
+    successCallback();
+  };
+}
 
 export function merge<Success, Error>(
   mapError: (error: NativeError) => Error,