+++ packages/react-native-core/src/internal/__tests__/Callback.test.ts
@@ -1,0 +1,16 @@
+import { mergeVoid } from '../index';
+
+describe('mergeVoid', () => {
+  it('should call success', () => {
+    // Given
+    let calledSuccess = false
+    // When
+    mergeVoid(
+      () => undefined,
+      () => {},
+      () => calledSuccess = true
+    )(undefined)
+    // Then
+    expect(calledSuccess).toBeTruthy()
+  })
+})