+++ packages/react-native-messaging-ui/ios/NablaMessagingUI/NablaMessagingUIModule.m
@@ -2,7 +2,7 @@
 
 @interface RCT_EXTERN_MODULE(NablaMessagingUIModule, NSObject)
 
-RCT_EXTERN_METHOD(navigateToInbox)
+RCT_EXTERN_METHOD(navigateToInbox: (RCTResponseSenderBlock)callback)
 RCT_EXTERN_METHOD(navigateToConversation: (id)conversationIdMap
                             showComposer: (BOOL)showComposer
                                 callback: (RCTResponseSenderBlock)callback)