+++ packages/react-native-messaging-ui/src/NablaMessagingUI/NablaMessagingUI.ts
@@ -1,7 +1,7 @@
 import { nablaMessagingUIModule } from './NablaMessagingUIModule';
 import { ConversationId } from '@nabla/react-native-messaging-core';
 import { NablaError } from '@nabla/react-native-core';
-import { merge } from '@nabla/react-native-core/lib/internal';
+import { mergeVoid } from '@nabla/react-native-core/lib/internal';
 import { mapError } from '@nabla/react-native-messaging-core/lib/internal';
 
 /**
@@ -13,9 +13,10 @@ import { mapError } from '@nabla/react-native-messaging-core/lib/internal';
 export class NablaMessagingUI {
   /**
    * Launches the Inbox Screen.
+   * @param dismissCallback optional callback called when screen is dismissed.
    */
-  public static navigateToInbox() {
-    nablaMessagingUIModule.navigateToInbox();
+  public static navigateToInbox(dismissCallback: () => void = () => {}) {
+    nablaMessagingUIModule.navigateToInbox(dismissCallback);
   }
 
   /**
@@ -23,21 +24,21 @@ export class NablaMessagingUI {
    * @param conversationId the id of the conversation to display.
    * @param showComposer optional flag to show or hide the composer.
    * @param errorCallback optional callback called in case of error. (Bad conversationId)
-   * @param successCallback optional callback called when navigation succeeds.
+   * @param dismissCallback optional callback called when screen is dismissed.
    */
   public static navigateToConversation(
     conversationId: ConversationId,
     showComposer: boolean = true,
     errorCallback: (error: NablaError) => void = () => {},
-    successCallback: () => void = () => {},
+    dismissCallback: () => void = () => {},
   ) {
     nablaMessagingUIModule.navigateToConversation(
       conversationId,
       showComposer,
-      merge(
+      mergeVoid(
         mapError,
         errorCallback,
-        successCallback,
+        dismissCallback,
       ));
   }
 }